pyproject.toml
[tool.poetry]
name = "llama-cpp-openai"
version = "0.1.0"
description = ""
authors = ["blav <blav@actar.us>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
fastapi = "^0.105.0"
llama-cpp-python = "^0.2.22"
pydantic = "^2.5.2"
uvicorn = "^0.24.0.post1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
httpx = "^0.25.2"
pyautogen = "^0.2.2"

[tool.poetry.scripts]
generate-openai-model = "scripts.generate_openai_model:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
